refactor: refactoring in some of the changes are updated
