+++ frontend/src/constants/api.ts
@@ -1,5 +1,5 @@
-export const ERROR_MESSAGE = 'Something went wrong';
+const ERROR_MESSAGE = 'Something went wrong';
 
 const getVersion = 'version';
 
-export { getVersion };
+export { ERROR_MESSAGE, getVersion };
