+++ frontend/src/container/GridGraphLayout/Graph/FullView/index.metricsBuilder.tsx
@@ -63,7 +63,6 @@ function FullView({
 		{
 			queryKey,
 		},
-		isDependedDataLoaded,
 	);
 
 	const chartDataSet = useMemo(
