+++ frontend/src/pages/MetricsApplication/contants.ts
